crud dla tagów

+++ app/src/Controller/CategoryController.php
@@ -54,6 +54,7 @@ class CategoryController extends AbstractController
             throw $this->createNotFoundException();
         }
 
+
         return $this->render(
             'category/view.html.twig',
             ['category' => $category]
@@ -142,6 +143,15 @@ class CategoryController extends AbstractController
     )]
     public function delete(Request $request, Category $category): Response
     {
+        if (!$this->categoryService->canBeDeleted($category)) {
+            $this->addFlash(
+                'warning',
+                'nie można usunąć kategorii'
+            );
+
+            return $this->redirectToRoute('category_index');
+        }
+
         $form = $this->createForm(FormType::class, $category, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
